requirements.txt
qdrant-client==1.12.2
fastembed==0.5.0
pillow==10.4.0
gradio==5.9.1
python-dotenv==1.0.1